Added river turning into drain in IPhreeqc version of PHAST
git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqc/trunk@8034 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ IPhreeqc.cpp
@@ -366,7 +366,7 @@ int IPhreeqc::GetSelectedOutputColumnCount(void)const
 
 int IPhreeqc::GetSelectedOutputCount(void)const
 {
-	return this->PhreeqcPtr->SelectedOutput_map.size();
+	return (int) this->PhreeqcPtr->SelectedOutput_map.size();
 }
 
 const char* IPhreeqc::GetSelectedOutputFileName(void)const
@@ -425,7 +425,7 @@ int IPhreeqc::GetSelectedOutputStringLineCount(void)const
 	std::map< int, std::vector < std::string > >::const_iterator cit = this->SelectedOutputLinesMap.find(this->CurrentSelectedOutputUserNumber);
 	if (cit != this->SelectedOutputLinesMap.end())
 	{
-		return (*cit).second.size();
+		return (int) (*cit).second.size();
 	}
 	return (int)0;
 }
